# SDT marker sdt_libc:memory_mallopt_mmap_max not available