Added library libc.so.6 to make memory_mallopt_mmap_max available
and added -o option for perf record to create an output file.
we should not create static/hard coded files.so,
created temp file and deleted at the end

Signed-off-by: shirisha Ganta <shiganta@in.ibm.com>