@shirishaganta 2020-05-12
Before Patch:
[root@zzfp219-lp1 perf]# avocado run perf_sdt_probe_org.py
/usr/local/lib/python3.6/site-packages/avocado_framework-77.0-py3.6.egg/avocado/plugins/run.py:287: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast, keep-tmp and ignore-missing-references.
  FutureWarning)
JOB ID     : 427a807b9add81816e59eec4eac23e268feee6c9
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-05-11T15.55-427a807/job.log
 (1/1) perf_sdt_probe_org.py:PerfSDT.test: FAIL: SDT marker sdt_libc:memory_mallopt_mmap_max not available (5.85 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 5.93 s
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-05-11T15.55-427a807/results.html

After Patch:
[root@zzfp219-lp1 perf]# avocado run perf_sdt_probe.py
/usr/local/lib/python3.6/site-packages/avocado_framework-77.0-py3.6.egg/avocado/plugins/run.py:287: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast, keep-tmp and ignore-missing-references.
  FutureWarning)
JOB ID     : 510256735d7222df098328c3c9154a1ee2802021
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-05-11T15.55-5102567/job.log
 (1/1) perf_sdt_probe.py:PerfSDT.test: PASS (8.51 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 8.58 s
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-05-11T15.55-5102567/results.html